separate config from state
