+++ protocol/v2/qbft/testing_utils.go
@@ -17,6 +17,7 @@ var TestingMessage = &specqbft.Message{
 var testingSignedMsg = func() *specqbft.SignedMessage {
 	return SignMsg(TestingSK, 1, TestingMessage)
 }()
+
 var SignMsg = func(sk *bls.SecretKey, id types.OperatorID, msg *specqbft.Message) *specqbft.SignedMessage {
 	domain := types.PrimusTestnet
 	sigType := types.QBFTSignatureType
@@ -30,12 +31,14 @@ var SignMsg = func(sk *bls.SecretKey, id types.OperatorID, msg *specqbft.Message
 		Signature: sig.Serialize(),
 	}
 }
+
 var TestingSK = func() *bls.SecretKey {
 	types.InitBLS()
 	ret := &bls.SecretKey{}
 	ret.SetByCSPRNG()
 	return ret
 }()
+
 var testingValidatorPK = spec.BLSPubKey{1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4}
 var testingShare = &types.Share{
 	OperatorID:      1,
@@ -51,6 +54,7 @@ var testingShare = &types.Share{
 		},
 	},
 }
+
 var TestingInstanceStruct = &specqbft.Instance{
 	State: &specqbft.State{
 		Share:                           testingShare,
@@ -93,6 +97,7 @@ var TestingInstanceStruct = &specqbft.Instance{
 		},
 	},
 }
+
 var TestingControllerStruct = &specqbft.Controller{
 	Identifier: []byte{1, 2, 3, 4},
 	Height:     specqbft.Height(1),
